+++ mistralrs-core/src/pipeline/loaders/vision_loaders.rs
@@ -20,6 +20,7 @@ use crate::pipeline::text_models_inputs_processor::{FlashParams, PagedAttentionI
 use crate::pipeline::{EitherCache, IsqModel, Processor, ProcessorCreator, VisionPromptPrefixer};
 use crate::vision_models::idefics2::{Config as Idefics2Config, Idefics2};
 use crate::vision_models::idefics2_input_processor::Idefics2Processor;
+use crate::vision_models::idefics3::{Idefics3Config, Idefics3Model, Idefics3Processor};
 use crate::vision_models::llava::config::Config as LLaVAConfig;
 use crate::vision_models::llava15::Model as LLaVA;
 use crate::vision_models::llava_inputs_processor::LLaVAProcessor;
@@ -94,6 +95,8 @@ pub enum VisionLoaderType {
     VLlama,
     #[serde(rename = "qwen2vl")]
     Qwen2VL,
+    #[serde(rename = "idefics3")]
+    Idefics3,
 }
 
 impl FromStr for VisionLoaderType {
@@ -106,7 +109,8 @@ impl FromStr for VisionLoaderType {
             "llava" => Ok(Self::LLaVA),
             "vllama" => Ok(Self::VLlama),
             "qwen2vl" => Ok(Self::Qwen2VL),
-            a => Err(format!("Unknown architecture `{a}`. Possible architectures: `phi3v`, `idefics2`, `llava_next`, `llava`, `vllama`, `qwen2vl`.")),
+            "idefics3" | "smolvlm" => Ok(Self::Idefics3),
+            a => Err(format!("Unknown architecture `{a}`. Possible architectures: `phi3v`, `idefics2`, `llava_next`, `llava`, `vllama`, `qwen2vl`, `idefics3`, `smolvlm`.")),
         }
     }
 }
@@ -601,5 +605,80 @@ impl IsqModelLoader for Qwen2VLLoader {
             Regex::new(r"layers\.(\d+)\.mlp\.up_proj\.(weight|bias)$")?,
             Regex::new(r"layers\.(\d+)\.mlp\.down_proj\.(weight|bias)$")?,
         ])
+    }
+}
+
+// ======================== Idefics 3 loader
+
+/// [`VisionLoader`] for an Idefics 3 Vision model.
+///
+/// [`VisionLoader`]: https://ericlbuehler.github.io/mistral.rs/mistralrs/struct.VisionLoader.html
+pub struct Idefics3Loader;
+
+pub struct Idefics3Prefixer;
+
+impl VisionPromptPrefixer for Idefics3Prefixer {
+    fn prefix_image(&self, _image_index: usize, prompt: &str) -> String {
+        // Chat template does it
+        format!("<image>{prompt}")
+    }
+}
+
+impl VisionModelLoader for Idefics3Loader {
+    fn load(
+        &self,
+        config: &str,
+        use_flash_attn: bool,
+        vb: VarBuilder,
+        normal_loading_metadata: NormalLoadingMetadata,
+        attention_mechanism: AttentionImplementation,
+    ) -> Result<Box<dyn VisionModel + Send + Sync>> {
+        let mut config: Idefics3Config = serde_json::from_str(config)?;
+        config.text_config.use_flash_attn = use_flash_attn;
+        Ok(Box::new(Idefics3Model::new(
+            &config,
+            vb,
+            self.is_gptx(),
+            normal_loading_metadata,
+            attention_mechanism,
+        )?))
+    }
+    fn is_gptx(&self) -> bool {
+        true
+    }
+    fn get_config_repr(&self, config: &str, use_flash_attn: bool) -> Result<Box<dyn Debug>> {
+        let mut config: Idefics3Config = serde_json::from_str(config)?;
+        config.text_config.use_flash_attn = use_flash_attn;
+        Ok(Box::new(config))
+    }
+    fn get_processor(
+        &self,
+        _model_config: &str,
+        processor_config: Option<ProcessorConfig>,
+        preprocessor_config: PreProcessorConfig,
+        max_edge: Option<u32>,
+    ) -> Arc<dyn Processor + Send + Sync> {
+        Arc::new(Idefics3Processor::new(
+            processor_config.unwrap(),
+            preprocessor_config,
+            max_edge,
+        ))
+    }
+    fn get_total_device_mapping_num_layers(&self, config: &str) -> Result<usize> {
+        let config: Idefics3Config = serde_json::from_str(config)?;
+        // We only apply device mapping to text model
+        Ok(config.text_config.num_hidden_layers)
+    }
+    fn supports_paged_attention(&self) -> bool {
+        true
+    }
+    fn prefixer(&self) -> Arc<dyn VisionPromptPrefixer> {
+        Arc::new(Idefics2Prefixer)
+    }
+}
+
+impl IsqModelLoader for Idefics3Loader {
+    fn isq_layer_regexes(&self, _config: &str) -> Result<Vec<Regex>> {
+        todo!()
     }
 }
